iOS 26 UX adjustments and general improvements in Map and Settings. Uploads can now continue when the app is in the background.

+++ CommonsFinder/Observable Models/AccountModel.swift
@@ -67,10 +67,11 @@ final class AccountModel {
 
     func logout() throws {
         postLoginTask?.cancel()
+        postLoginTask = nil
+        recurringSyncTask?.cancel()
+        recurringSyncTask = nil
         activeUser = nil
-        let deletedCount = try appDatabase.deleteAllImageModels()
-        // FIXME: also delete categories and all interactions
-        logger.info("Deleted \(deletedCount) items on-logout")
+        try appDatabase.deleteLogoutRelatedItems()
         try Authentication.clearKeychain()
     }
 
@@ -187,6 +188,9 @@ final class AccountModel {
             try appDatabase
             .fetchAllFiles(byUsername: username, withNames: draftFinalFilenames)
             .map(\.name)
+            .filter { !$0.isEmpty }
+
+        guard !draftsToCleanup.isEmpty else { return }
 
         let deletedFileCount = try appDatabase.deleteDrafts(withFinalFilenames: draftsToCleanup)
         if deletedFileCount != 0 {
